requirements.txt
# Core FastAPI Dependencies
fastapi==0.104.1
pydantic==2.4.2
uvicorn[standard]==0.24.0

# HTTP client for SharePoint integration  
requests==2.31.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For FastAPI TestClient

# Development dependencies
black==23.10.1
flake8==6.1.0
python-dotenv==1.0.0

# File handling and validation
python-magic==0.4.27
Pillow>=10.0.0

# Production server
gunicorn==21.2.0

# Configuration management
PyYAML==6.0.1
